Move key-value related AST to dedicated namespace (PHRAS-688)
