+++ lib/Alchemy/Phrasea/SearchEngine/Elastic/AST/KeyValue/Key.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Alchemy\Phrasea\SearchEngine\Elastic\AST\KeyValue;
+
+interface Key
+{
+    public function getIndexField();
+    public function __toString();
+}
