+++ lib/Alchemy/Phrasea/SearchEngine/Elastic/Search/QueryVisitor.php
@@ -126,8 +126,8 @@ class QueryVisitor implements Visit
         $field = $this->visit($element->getChild(1));
         if ($field instanceof AST\Field) {
             return new AST\InExpression($field, $this->visit($expression));
-        } elseif ($field instanceof AST\Key) {
-            return new AST\KeyValueExpression(
+        } elseif ($field instanceof AST\KeyValue\Key) {
+            return new AST\KeyValue\Expression(
                 $field,
                 $this->visitString($expression)
             );
@@ -319,13 +319,13 @@ class QueryVisitor implements Visit
         $type = $element->getChild(0)->getValue()['token'];
         switch ($type) {
             case NodeTypes::TOKEN_DATABASE:
-                return AST\Key::database();
+                return AST\KeyValue\NativeKey::database();
             case NodeTypes::TOKEN_COLLECTION:
-                return AST\Key::collection();
+                return AST\KeyValue\NativeKey::collection();
             case NodeTypes::TOKEN_MEDIA_TYPE:
-                return AST\Key::mediaType();
+                return AST\KeyValue\NativeKey::mediaType();
             case NodeTypes::TOKEN_RECORD_ID:
-                return AST\Key::recordIdentifier();
+                return AST\KeyValue\NativeKey::recordIdentifier();
             default:
                 throw new InvalidArgumentException(sprintf('Unexpected token type "%s" for native key.', $type));
         }
@@ -338,7 +338,7 @@ class QueryVisitor implements Visit
         }
         $baseName = $element->getChild(0)->getValue()['value'];
 
-        return new AST\KeyValueExpression(AST\Key::database(), $baseName);
+        return new AST\KeyValue\Expression(AST\KeyValue\NativeKey::database(), $baseName);
     }
 
     private function visitCollectionNode(Element $element)
@@ -348,7 +348,7 @@ class QueryVisitor implements Visit
         }
         $collectionName = $element->getChild(0)->getValue()['value'];
 
-        return new AST\KeyValueExpression(AST\Key::collection(), $collectionName);
+        return new AST\KeyValue\Expression(AST\KeyValue\NativeKey::collection(), $collectionName);
     }
 
     private function visitTypeNode(Element $element)
@@ -358,7 +358,7 @@ class QueryVisitor implements Visit
         }
         $typeName = $element->getChild(0)->getValue()['value'];
 
-        return new AST\KeyValueExpression(AST\Key::mediaType(), $typeName);
+        return new AST\KeyValue\Expression(AST\KeyValue\NativeKey::mediaType(), $typeName);
     }
 
     private function visitIdentifierNode(Element $element)
@@ -368,6 +368,6 @@ class QueryVisitor implements Visit
         }
         $identifier = $element->getChild(0)->getValue()['value'];
 
-        return new AST\KeyValueExpression(AST\Key::recordIdentifier(), $identifier);
+        return new AST\KeyValue\Expression(AST\KeyValue\NativeKey::recordIdentifier(), $identifier);
     }
 }
